Separate config file for panda_controller in demos
Demos use a different action space than traditional environments. Their primary purpose is to evaluate controller performance for small deltas. The input/output scaling for controllers should there for be maintained at 1.0, so that the delta_val may be adjusted in the demo script, rather than on the config side.