Fixing examples
Changes to config specification resulted in changes for the example folder. Switch Simple Reach
removed as it was redundant with Simple Reach and not entirely functional. Simple Reach behavior for Real Robots was also modified to adjust goal positions from robot's position after resetting. This because the space for goal positions that might result in collisions vary from robot to robot. The Sawyer has a much larger reach than the Panda. Therefore, it is safer to set these positions from the robot's reset position.