Change redis conf from auth to bind
Change in redis-server conf for franka-panda from require authentication via password to binding an ip address on the control pc. This only allows clients to connect to the server locally
(on the same machine) or on the local ethernet network. This simplifies setup and avoids forcing perls2 scripts to have access to the redis_passfile